+++ src/linear_objects/dike_traject.py
@@ -26,6 +26,7 @@ class DikeTraject(BaseLinearObject):
     _run_id_vr: int
     _run_id_dsn: int
     run_name: str = None
+    final_step_number: int = None  # the step number of the final step in the greedy optimization
 
     def serialize(self) -> dict:
         """Serialize the DikeTraject object to a dict, in order to be saved in dcc.Store"""
@@ -40,6 +41,7 @@ class DikeTraject(BaseLinearObject):
             "run_name": self.run_name,
             "_run_id_vr": self._run_id_vr,
             "_run_id_dsn": self._run_id_dsn,
+            "final_step_number": self.final_step_number,
         }
 
     @staticmethod
@@ -63,6 +65,7 @@ class DikeTraject(BaseLinearObject):
             run_name=data["run_name"],
             _run_id_vr=data["_run_id_vr"],
             _run_id_dsn=data["_run_id_dsn"],
+            final_step_number=data["final_step_number"],
         )
 
     def export_to_geojson(self, params: dict) -> str:
@@ -84,6 +87,12 @@ class DikeTraject(BaseLinearObject):
         return json.dumps(_geojson)
 
     def calc_traject_probability_array(self, calc_type: str) -> np.array:
+        """
+        Return an array of the traject probability of failure for year year and each step. Columns are the years and
+        rows are the steps. The first row is the probability of failure of the unreinforced dike traject.
+        :param calc_type:
+        :return:
+        """
 
         _beta_df = get_initial_assessment_df(self.dike_sections)
         _traject_pf, _ = get_traject_prob(_beta_df)
@@ -124,24 +133,15 @@ class DikeTraject(BaseLinearObject):
                         _beta_df["mechanism"] == mechanism
                 )
                 # replace the row in the dataframe with the betas of the section if both the name and mechanism match
-                d = {
-                    "name": section.name,
-                    "mechanism": mechanism,
-                    "Length": section.length,
-                }
-
-
 
                 for year, beta in zip(
                         years, getattr(section, _section_measure)[mechanism]
                 ):
-                    d[year] = beta
-                _beta_df.loc[mask, years] = d
+                    _beta_df.loc[mask, year] = beta
 
             _reinforced_traject_pf, _ = get_traject_prob(_beta_df)
 
             _traject_pf = np.concatenate((_traject_pf, _reinforced_traject_pf), axis=0)
-
         return np.array(_traject_pf)
 
     def get_section(self, name: str) -> DikeSection: